+++ internal/cmd/x.go
@@ -1,43 +1,39 @@
 package cmd
 
 import (
-	"errors"
 	"fmt"
-	"os"
-	"os/exec"
-	"path/filepath"
 
 	"github.com/spf13/cobra"
 	"github.com/spf13/viper"
 
 	"github.com/craftcms/nitro/config"
+	"github.com/craftcms/nitro/internal/find"
 )
 
 var xCommand = &cobra.Command{
 	Use:    "x",
 	Hidden: true,
 	RunE: func(cmd *cobra.Command, args []string) error {
-		cfgFile := viper.ConfigFileUsed()
-		if cfgFile == "" {
-			return errors.New("unable to find the config file")
+		machine := "nitro-dev"
+		if flagMachineName != "" {
+			machine = flagMachineName
 		}
 
-		filePath, err := filepath.Abs(cfgFile)
+		var cfg config.Config
+		if err := viper.Unmarshal(&cfg); err != nil {
+			return err
+		}
+
+		dbs, err := find.ContainersToCreate(machine, cfg)
 		if err != nil {
 			return err
 		}
 
-		nitro, err := exec.LookPath("nitro")
-		if err != nil {
-			return err
+		for _, db := range dbs {
+			fmt.Println(db.Engine)
 		}
 
-		fmt.Println("Modifying the hosts file to add sites for", config.GetString("name", ""), "(you will be prompted for your password)... ")
-		hostsCmd := exec.Command("sudo", nitro, "-f", filePath, "hosts", "add")
-		hostsCmd.Stdout = os.Stdout
-		hostsCmd.Stderr = os.Stderr
-
-		return hostsCmd.Run()
+		return nil
 	},
 }